refactor: split LLM providers and isolate provider options
