+++ tests/test_config.py
@@ -17,9 +17,15 @@ class TestConfigFileCreation(unittest.TestCase):
 
             self.assertTrue(path.is_file())
             self.assertEqual(cfg.llm.provider, "deepseek")
-            self.assertIn("strong", cfg.llm.tiers)
+            self.assertIsNone(cfg.llm.base_url)
+            self.assertIsNone(cfg.llm.api_key_env)
+            self.assertEqual(cfg.llm.tiers, {})
+            self.assertFalse(hasattr(cfg.llm, "api_key"))
             generated = path.read_text(encoding="utf-8")
             self.assertIn("# trans-novel 配置", generated)
+            self.assertNotIn("  base_url:", generated)
+            self.assertNotIn("  api_key_env:", generated)
+            self.assertNotIn("  tiers:", generated)
             self.assertIn("output:\n", generated)
             self.assertTrue(cfg.output.mono)
             self.assertFalse(cfg.output.bilingual)
